requirements.txt
attrs==16.2.0
certifi==2016.9.26
cffi==1.8.3
click==6.6
constantly==15.1.0
cryptography==1.5.2
cssselect==1.0.0
Flask==0.11.1
future==0.16.0
idna==2.1
incremental==16.10.1
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.4
MarkupSafe==0.23
parsel==1.0.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
python-telegram-bot==5.2.0
queuelib==1.4.2
Scrapy==1.2.1
service-identity==16.0.0
six==1.10.0
Twisted==16.5.0
urllib3==1.19
w3lib==1.15.0
Werkzeug==0.11.11
zope.interface==4.3.2
